% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Inventory_Item.R
\name{Add_Inventory_Item}
\alias{Add_Inventory_Item}
\title{Adds Inventory Item to Database}
\usage{
Add_Inventory_Item(inv_record)
}
\arguments{
\item{inv_record}{The inventory items to add to the database.}
}
\value{
Updated Inventory Table as data frame.
}
\description{
This function Adds Inventory Item to Database.
}
\details{
Requires: RPostgres, dplyr, shinytoastr
}
\examples{

}
