% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Append_Table.R
\name{DB_Append_Table}
\alias{DB_Append_Table}
\title{Appends a Record or Records to a Database Table}
\usage{
DB_Append_Table(table_name, table_record)
}
\arguments{
\item{table_name}{The name of the DB Table to check.}

\item{table_record}{The record(s) to write to the Database Table.}
}
\value{

}
\description{
This function appends to a Database Table.
}
\details{
Requires: RPostgres
}
