% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Connect.R
\name{DB_Connect}
\alias{DB_Connect}
\title{Connect to PostgreSQL Database}
\usage{
DB_Connect(db_name, db_host, db_user, db_pass)
}
\arguments{
\item{db_name}{The name of the database.}

\item{db_host}{The database hostname.}

\item{db_user}{The database user ID.}

\item{db_pass}{The database password for the user ID.}
}
\value{
db_con The connection parameters to the database.
}
\description{
This function connects to the PostgreSQL Database.
}
\details{
Requires: RPostgres
}
\examples{

}
