% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Create_Table.R
\name{DB_Create_Table}
\alias{DB_Create_Table}
\title{Creates a Database Table}
\usage{
DB_Create_Table(db_con, db_create_name, table_create)
}
\arguments{
\item{db_con}{The DB Connection parameters.}

\item{db_create_name}{The name of the DB Table to create.}

\item{table_create}{The data frame to use as the source for the new Table.}
}
\value{

}
\description{
This function Creates a Database Table.
}
\details{
Requires: RPostgres
}
