% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Disconnect.R
\name{DB_Disconnect}
\alias{DB_Disconnect}
\title{Disconnect the PostgreSQL Database}
\usage{
DB_Disconnect(db_con)
}
\arguments{
\item{db_con}{The connection parameters to the database.}
}
\value{
db_con The connection parameters to the database.
}
\description{
This function disconnects the PostgreSQL Database.
}
\details{
Requires: RPostgres
}
