% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Execute_Command.R
\name{DB_Execute_Command}
\alias{DB_Execute_Command}
\title{Executes a Command to the Database}
\usage{
DB_Execute_Command(my_command)
}
\arguments{
\item{my_command}{The command to execute.}
}
\description{
This function Executes a Command to the Database.
}
\details{
Requires: RPostgres
}
