% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Inventory_Record_Update.R
\name{DB_Inventory_Record_Update}
\alias{DB_Inventory_Record_Update}
\title{Updates Inventory Record(s) in Database}
\usage{
DB_Inventory_Record_Update(inv_record, changed_idx, t_action)
}
\arguments{
\item{inv_record}{The inventory items to update in the database.}

\item{changed_idx}{The index to changed fields.}

\item{t_action}{The type of update action being performed.}
}
\value{
Updated Inventory Table as data frame.
}
\description{
This function Updates Inventory Record(s) in Database.
}
\details{
Requires: RPostgres, dplyr, shinytoastr
}
