% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_List_Fields.R
\name{DB_List_Fields}
\alias{DB_List_Fields}
\title{List Fields in Database Table}
\usage{
DB_List_Fields(db_con, table_name)
}
\arguments{
\item{db_con}{The DB Connection parameters.}

\item{table_name}{The name of the DB Table to check.}
}
\value{
field_list Character array of field names.
}
\description{
This function lists the fields in a database table.
}
\details{
Requires: RPostgres
}
