% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_List_Tables.R
\name{DB_List_Tables}
\alias{DB_List_Tables}
\title{List Tables in Database}
\usage{
DB_List_Tables(db_con)
}
\arguments{
\item{db_con}{The DB Connection parameters.}
}
\value{
table_list Character array.
}
\description{
This function lists the tables in the database.
}
\details{
Requires: RPostgres
}
