% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Read_Table.R
\name{DB_Read_Table}
\alias{DB_Read_Table}
\title{Reads a Database Table}
\usage{
DB_Read_Table(table_name)
}
\arguments{
\item{table_name}{The name of the DB Table to check.}
}
\value{
query_table Data frame containing the table data.
}
\description{
This function Reads a Database Table.
}
\details{
Requires: RPostgres
}
