% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Read_Table_User.R
\name{DB_Read_Table_User}
\alias{DB_Read_Table_User}
\title{Queries Database and Returns Specified Table Rows for User}
\usage{
DB_Read_Table_User(table_name, db_user_id = NULL)
}
\arguments{
\item{table_name}{The table name to perform query.}

\item{db_user_id}{The user ID for the table query.}
}
\value{
query_table The returned table for user.
}
\description{
This function Queries Database and Returns Specified Table Rows for User
}
\details{
Requires: RPostgres
}
