% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Remove_Table.R
\name{DB_Remove_Table}
\alias{DB_Remove_Table}
\title{Removes a Database Table}
\usage{
DB_Remove_Table(db_con, table_name)
}
\arguments{
\item{db_con}{The DB Connection parameters.}

\item{table_name}{The name of the DB Table to remove.}
}
\value{

}
\description{
This function Removes a Database Table.
}
\details{
Requires: RPostgres
}
