% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Table_Exists.R
\name{DB_Table_Exists}
\alias{DB_Table_Exists}
\title{Check for Table in Database}
\usage{
DB_Table_Exists(db_con, table_name)
}
\arguments{
\item{db_con}{The DB Connection parameters.}

\item{table_name}{The name of the DB Table to check.}
}
\value{
table_exist Logical value.
}
\description{
This function checks for the existence of a specific table in the database.
}
\details{
Requires: RPostgres
}
\examples{

}
