% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Inventory_User.R
\name{Get_Inventory_User}
\alias{Get_Inventory_User}
\title{Queries Database and Returns Inventory Table}
\usage{
Get_Inventory_User()
}
\value{
Inventory Table as a data frame.
}
\description{
This function Queries the Database and Returns an Inventory Table for the user.
}
\details{
Requires: RPostgres, dplyr
}
\examples{

}
