% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_PM_Sales_Activity_User.R
\name{Get_PM_Sales_Activity_User}
\alias{Get_PM_Sales_Activity_User}
\title{Queries Database for Poshmark Sales Activity}
\usage{
Get_PM_Sales_Activity_User()
}
\value{
Poshmark Sales Table as data frame.
}
\description{
This function Queries Database for Poshmark Sales Activity for user.
}
\details{
Requires: RPostgres, dplyr
}
\examples{

}
