% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_User_Goals.R
\name{Get_User_Goals}
\alias{Get_User_Goals}
\title{Queries Database for User Goals}
\usage{
Get_User_Goals()
}
\value{
User Goals Table as data frame.
}
\description{
This function Queries Database for User Goals.
}
\details{
Requires: RPostgres, dplyr
}
\examples{

}
