% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_User_Record.R
\name{Get_User_Record}
\alias{Get_User_Record}
\title{Retrieve User Record from User Table}
\usage{
Get_User_Record(t_user)
}
\arguments{
\item{t_user}{The user ID to retrieve the record for.}
}
\value{
user_record The user record in the table.
}
\description{
This function Fetches the User Information for Logon and Authorization.
}
\details{
Requires: RPostgres
}
\examples{

}
