% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Process_Return.R
\name{PM_Process_Return}
\alias{PM_Process_Return}
\title{Deletes Poshmark Sales Activity Record from Database}
\usage{
PM_Process_Return(inv_record)
}
\arguments{
\item{inv_record}{}
}
\value{
Poshmark Sales Table as data frame.
}
\description{
This function Deletes Poshmark Sales Activity Record from Database.
}
\details{
Requires: RPostgres
}
\examples{

}
