% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Inventory_Item.R
\name{Update_Inventory_Item}
\alias{Update_Inventory_Item}
\title{Updates Inventory Item(s) in Database}
\usage{
Update_Inventory_Item(inv_record, changed_idx, table_names, t_action)
}
\arguments{
\item{inv_record}{The inventory items to update in the database.}
}
\value{
Updated Inventory Table as data frame.
}
\description{
This function Updates Inventory Item(s) in Database.
}
\details{
Requires: RPostgres, dplyr, shinytoastr
}
\examples{

}
