% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_User_Goals.R
\name{Update_User_Goals}
\alias{Update_User_Goals}
\title{Updates User Goals in Database}
\usage{
Update_User_Goals(u_user_goals, changed_tbl, table_names)
}
\arguments{
\item{u_user_goals}{}
}
\value{
User Goals Table as data frame.
}
\description{
This function Updates User Goals in Database.
}
\details{
Requires: RPostgres, dplyr
}
\examples{

}
