% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_User_Last.R
\name{Update_User_Last}
\alias{Update_User_Last}
\title{Update User Record with Last Logon Timestamp}
\usage{
Update_User_Last(user_record)
}
\arguments{
\item{user_record}{The User Record to update.}
}
\value{
session_user_record Updated session information.
}
\description{
This function Updates User Record with Last Logon Timestamp.
}
\details{
Requires: RPostgres
}
\examples{

}
