% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_User_Password.R
\name{Update_User_Password}
\alias{Update_User_Password}
\title{Update User Record with New Password}
\usage{
Update_User_Password(pwd_hash)
}
\arguments{
\item{pwd_hash}{The hashed password for the user.}
}
\value{
session_user_record Updated session record.
}
\description{
This function Updates User Record with new password.
}
\details{
Requires: RPostgres
}
\examples{

}
