% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_User_Profile.R
\name{Update_User_Profile}
\alias{Update_User_Profile}
\title{Update User Profile}
\usage{
Update_User_Profile(user_record, changed_idx)
}
\arguments{
\item{user_record}{The user record.}

\item{changed_idx}{The index of changed fields.}
}
\value{
session_user_record
}
\description{
This function Updates User Profile information in User Table.
}
\details{
Requires: RPostgres
}
\examples{

}
