% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_Sales_Activity_User.R
\name{Write_Sales_Activity_User}
\alias{Write_Sales_Activity_User}
\title{Writes Poshmark Sales Activity to Database}
\usage{
Write_Sales_Activity_User(sales_activity)
}
\arguments{
\item{sales_activity}{}
}
\value{
Poshmark Sales Table as data frame.
}
\description{
This function Writes Poshmark Sales Activity to Database for user.
}
\details{
Requires: RPostgres, dplyr
}
\examples{

}
